refactor: 출력을 데이터만드는 분리
