+++ src/main/kotlin/lotto/Application.kt
@@ -1,13 +1,15 @@
 package lotto
 
 import lotto.controller.LottoController
+import lotto.util.RandomNumber
 import lotto.view.InputView
 import lotto.view.OutputView
 
-val input = InputView()
+val randomNumber = RandomNumber()
+val inputView = InputView()
 val outputView = OutputView()
 
 fun main() {
-    val lottoController = LottoController()
+    val lottoController = LottoController(inputView, randomNumber)
     lottoController.run()
 }
